Elasticmodify

+++ src/main/java/iudx/file/server/database/elasticdb/utilities/ResponseBuilder.java
@@ -52,7 +52,7 @@ public class ResponseBuilder {
     return this;
   }
 
-  public ResponseBuilder setCount(int count) {
+  public ResponseBuilder setCount(long count) {
     response.put(RESULTS, new JsonArray().add(new JsonObject().put(COUNT, count)));
     return this;
   }
